+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.5.0
+
+- draft skills
+- draft foraging
+
 # 0.4.0
 
 - update